Fix for Issue #9 - running without Info.plist
NSUserNotificationCenter.defaultNotificationCenter() returns None if there
is no CFBundleIdentifier in python's plist (MacOS app configuration) file.

The default system python has an associated plist; but pythons run in a
virtualenv do not.

This fix force-sets the CFBundleIdentifier field to be the application name
if it is not already set in the plist.